implenenting user auth

+++ src/easyway/easyway.js
@@ -4,12 +4,13 @@ const router = express.Router();
 const { Event, Person, schemaName } = require('../mongodb');
 const { authenticateToken } = require('../auth');
 const { asyncForEach } = require('../utils/functions');
+const { roles } = require('../userdb');
 
 /**
  * Get db Collection as array
  * Headers = {Collection = ""}
  */
-router.get('/collection', authenticateToken, async(req, res, next) => {
+router.get('/collection', authenticateToken, async (req, res, next) => {
     logger.info('fetch all ' + req.headers.collection + ' from db');
     try {
         //perhabs this could be done less often
@@ -27,7 +28,7 @@ router.get('/collection', authenticateToken, async(req, res, next) => {
  * Body = Entry object
  * Headers = {collection = ""}
  */
-router.post('/add', authenticateToken, async(req, res, next) => {
+router.post('/add', authenticateToken, async (req, res, next) => {
     logger.info(`add object to ${req.headers.collection}`);
     try {
         const model = getMongooseModel(req.headers.collection);
@@ -51,7 +52,7 @@ router.post('/add', authenticateToken, async(req, res, next) => {
  * Body = Object only with changed properties
  * Headers = {collection = ""}
  */
-router.put('/change/:id', authenticateToken, async(req, res, next) => {
+router.put('/change/:id', authenticateToken, async (req, res, next) => {
     logger.info(`change in ${req.headers.collection} this -> ${req.params.id}`);
     try {
         const model = getMongooseModel(req.headers.collection);
@@ -70,13 +71,17 @@ router.put('/change/:id', authenticateToken, async(req, res, next) => {
  * Delete mongodb entry
  * Headers = {collection = ""}
  */
-router.delete('/delete/:id', authenticateToken, async(req, res, next) => {
+router.delete('/delete/:id', authenticateToken, async (req, res, next) => {
     logger.info(`delete in collection ${req.headers.collection} this -> ${req.params.id}`);
     try {
-        const model = getMongooseModel(req.headers.collection);
-        await model.deleteOne({ _id: req.params.id });
-        await deleteDependendItems(req.params.id, req.headers.collection);
-        res.status(200).send();
+        if (req.user.roles === roles.ADMIN) {
+            const model = getMongooseModel(req.headers.collection);
+            await model.deleteOne({ _id: req.params.id });
+            await deleteDependendItems(req.params.id, req.headers.collection);
+            res.status(200).send();
+        } else {
+            res.status(401).send({ message: 'Permission denied' });
+        }
     } catch (err) {
         logger.error("Delete object failed: " + err.message);
         next(err);
@@ -94,7 +99,7 @@ function getMongooseModel(modelName) {
 async function recalCalculateAge() {
     const personModel = getMongooseModel(schemaName.PERSON);
     let persons = await personModel.find({});
-    asyncForEach(persons, async(personItem) => {
+    asyncForEach(persons, async (personItem) => {
         if (personItem.person.firstname !== "#DUMMY") {
             let ageDifMs = Date.now() - new Date(personItem.person.birthdate).getTime();
             let ageDate = new Date(ageDifMs);
@@ -111,14 +116,14 @@ async function deleteDependendItems(id, model) {
         if (model === schemaName.EVENT) {
             const personModel = getMongooseModel(schemaName.PERSON);
             let persons = await personModel.find({});
-            asyncForEach(persons, async(personItem) => {
+            asyncForEach(persons, async (personItem) => {
                 personItem.person.event = personItem.person.event.filter(item => item._id !== id);
                 await personModel.updateOne({ _id: personItem._id }, { $set: personItem });
             });
         } else {
             const eventModel = getMongooseModel(schemaName.EVENT);
             let events = await eventModel.find({});
-            asyncForEach(events, async(eventItem) => {
+            asyncForEach(events, async (eventItem) => {
                 eventItem.event.participants.splice(eventItem.event.participants.indexOf(id), 1);
                 await eventModel.updateOne({ _id: eventItem._id }, { $set: eventItem });
             });
@@ -132,7 +137,7 @@ async function refreshEventsDB(id, body) {
     try {
         const eventModel = getMongooseModel(schemaName.EVENT);
         let events = await eventModel.find({});
-        asyncForEach(events, async(element) => {
+        asyncForEach(events, async (element) => {
             if (isIncluded(element._id, body.person.event) && !element.event.participants.includes(id)) {
                 element.event.participants.push(id.toString());
             } else if (!isIncluded(element._id, body.person.event) && element.event.participants.includes(id)) {
